requirements.txt
agate==1.9.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.13.0
asyncio==4.0.0
attrs==26.1.0
babel==2.18.0
backoff==2.2.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
clickhouse-connect==0.15.1
clickhouse-driver==0.2.10
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==14.0
contourpy==1.3.3
cycler==0.12.1
daff==1.4.2
dagster==1.13.2
dagster-dbt==0.29.2
dagster-graphql==1.13.2
dagster-pipes==1.13.2
dagster-shared==1.13.2
dagster-webserver==1.13.2
dbt-adapters==1.22.5
dbt-clickhouse==1.10.0
dbt-common==1.37.5
dbt-core==1.11.8
dbt-extractor==0.6.0
dbt-protos==1.0.474
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.2
docstring_parser==0.18.0
fastapi==0.136.1
filelock==3.29.0
fonttools==4.62.1
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.47
gql==4.0.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.4.0
grpcio==1.80.0
grpcio-health-checking==1.80.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanfriendly==10.0
idna==3.13
importlib_metadata==8.9.0
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kafka-python==2.3.1
kiwisolver==1.5.0
leather==0.4.1
llvmlite==0.47.0
loguru==0.7.3
lz4==4.4.5
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mashumaro==3.14
matplotlib==3.10.9
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
numba==0.65.1
numpy==2.4.4
nvidia-nccl-cu12==2.30.4
orderly-set==5.5.0
orjson==3.11.8
packaging==26.2
pandas==3.0.2
parsedatetime==2.6
pathlib_abc==0.5.2
pathspec==0.12.1
pillow==12.2.0
platformdirs==4.9.6
propcache==0.4.1
protobuf==6.33.6
py4j==0.10.9.7
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pyparsing==3.3.2
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
shap==0.51.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smmap==5.0.3
snowplow-tracker==1.1.0
SQLAlchemy==2.0.49
sqlglot==28.0.0
sqlglotrs==0.7.3
sqlparse==0.5.5
starlette==1.0.0
structlog==25.5.0
tabulate==0.10.0
text-unidecode==1.3
threadpoolctl==3.6.0
tomli==2.4.1
tomlkit==0.14.0
toposort==1.10
tqdm==4.67.3
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
universal_pathlib==0.3.10
urllib3==2.6.3
uvicorn==0.46.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
xgboost==3.2.0
yarl==1.23.0
zipp==3.23.1
zstandard==0.25.0
